дополняет прошлое домашнее задание
